# back_end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MusicApp API",
        "description": "Documentation de l'API MusicApp (auth, artistes, albums, chansons.)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/albums": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Get all albums",
                "description": "Retrieve a list of all albums with their artist",
                "operationId": "1da83b13e3f02c4fee3c24e70098d603",
                "responses": {
                    "200": {
                        "description": "List of albums",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "titre": {
                                                "type": "string",
                                                "example": "Album 1"
                                            },
                                            "annee": {
                                                "type": "integer",
                                                "example": 2025
                                            },
                                            "artist_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Albums"
                ],
                "summary": "Create a new album",
                "description": "Store a newly created album",
                "operationId": "e2953334b8702ac4d6cc93c63b7440ef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titre",
                                    "annee",
                                    "artist_id"
                                ],
                                "properties": {
                                    "titre": {
                                        "type": "string",
                                        "example": "Album 1"
                                    },
                                    "annee": {
                                        "type": "integer",
                                        "example": 2025
                                    },
                                    "artist_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Album created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "titre": {
                                            "type": "string",
                                            "example": "Album 1"
                                        },
                                        "annee": {
                                            "type": "integer",
                                            "example": 2025
                                        },
                                        "artist_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/albums/{id}": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Get album details",
                "description": "Retrieve details of a specific album by ID",
                "operationId": "6a68209549d4d66cd13500611af64336",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Album ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Album details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "titre": {
                                            "type": "string",
                                            "example": "Album 1"
                                        },
                                        "annee": {
                                            "type": "integer",
                                            "example": 2025
                                        },
                                        "artist_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Album not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Albums"
                ],
                "summary": "Update an album",
                "description": "Update the details of an existing album by ID",
                "operationId": "9d45cc434d36c93ef067bbd2be12a53c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Album ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titre": {
                                        "type": "string",
                                        "example": "Updated Album"
                                    },
                                    "annee": {
                                        "type": "integer",
                                        "example": 2026
                                    },
                                    "artist_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Album updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Album updated successfully"
                                        },
                                        "album": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Album not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Albums"
                ],
                "summary": "Delete an album",
                "description": "Delete an existing album by ID",
                "operationId": "b148d53b889a5c777b217bb73821307e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Album ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Album deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Album deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Album not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use your API token from /login",
                "bearerFormat": "API Token",
                "scheme": "bearer"
            }
        }
    },
...
}